+++ CHANGELOG.md
@@ -2,10 +2,10 @@
 ## Update Notes:
 - CPR for v14 does not contain all automations that previously existed in V13.
 - Requires CAT version X.Y.Z.
-- Added Generic: damageBonusToOneRoll, passiveAura, hide, correctSpellLink
+- Added Generic: damageBonusToOneRoll, passiveAura, hide, correctSpellLink, pushOnFail
 ## Bug Fixes:
 - Brutal Strike
-- Generic: multiSingleTarget
+- Generic: multiSingleTarget, proneOnHit, proneOnFail
 - Empowered Evocation
 - Song of Defense
 - Greater Divine Intervention
@@ -20,6 +20,15 @@
 - Wild Shape
 - Wild Companion
 - Beast Spells
+- Archery
+- Defense
+- Dueling
+- Great Weapon Fighting
+- Interception
+- Protection
+- Thrown Weapon Fighting
+- Unarmed Fighting
+- Unarmed Strike
 ## Updated Modern Automations:
 - Channel Divinity (Cleric)
 - Sear Undead
@@ -33,6 +42,15 @@
 - Wild Companion
 - Elemental Fury
 - Beast Spells
+- Archery
+- Defense
+- Dueling
+- Great Weapon Fighting
+- Interception
+- Protection
+- Thrown Weapon Fighting
+- Unarmed Fighthing
+- Unarmed Strike
 ## New Modern Automations:
 - Hand of Death